Repaired memory leaks (CTPRecord.cpp). Removed non-standard qualifications (citydata.h). Various cleanups.
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@690 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/net/general/net_action.cpp
@@ -321,7 +321,7 @@ void NetAction::Packetize(uint8* buf, uint16& size)
 	buf[1] = 'A';
 	size = 2;
 
-	PUSHSHORT(m_action);
+	PUSHSHORT(static_cast<uint16>(m_action));
 	for(uint32 i = 0; i < m_args[m_action]; i++) {
 		PUSHLONG(m_data[i]);
 	}
